fixed delete bug on BarcodeScanner

+++ src/screens/BarcodeScanner.js
@@ -62,29 +62,41 @@ const setBarcodeTimer = setBarcode => {
   }, 2000);
 };
 
-const addBook = async (realm, item, setList) => {
+const addBook = async ({realm, item, setList, setError}) => {
   const category = await Database.saveCategoryName(realm, item.categoryName);
   console.log(category.id, category.parentId, category.name, category.level);
   const toc = item.bookinfo && item.bookinfo.toc;
   const book = await Database.saveBook(realm, {...item, ...{category, toc}});
   console.log('new book', book.id, book.title);
-  item._alreadyAdded = book !== null;
-  setList([item]);
+  if (book) {
+    book._alreadyAdded = true;
+    setList([book]);
+  } else {
+    item._alreadyAdded = false;
+    setList([item]);
+  }
+  setError(null);
 };
 
-const deleteBook = (realm, item, setList) => {
-  const book = {...item};
+const deleteBook = ({realm, item, setList, setError}) => {
+  const bookId = item.id;
+  console.log('Database.deleteBookById bookId', bookId);
+  if (!bookId) {
+    console.log('Database.deleteBookById bookId is not defined');
+    return;
+  }
   Database.deleteBookById(realm, item.id)
     .then(() => {
-      console.log('Database.deleteBookById done', book.id, book.title);
+      console.log('Database.deleteBookById done', bookId);
       setList([]);
+      setError(null);
     })
     .catch(e => {
-      console.log('Database.deleteBookById error', book.id, e);
+      console.log('Database.deleteBookById error', bookId, e);
     });
 };
 
-const getIcon = (realm, item, setList) => {
+const getIcon = ({realm, item, setList, setError}) => {
   if (!item._alreadyAdded) {
     return (
       <Icon
@@ -92,7 +104,7 @@ const getIcon = (realm, item, setList) => {
         name="add"
         type="material"
         onPress={() => {
-          addBook(realm, item, setList);
+          addBook({realm, item, setList, setError});
         }}
       />
     );
@@ -102,15 +114,16 @@ const getIcon = (realm, item, setList) => {
         reverse
         name="delete"
         type="material"
+        color="crimson"
         onPress={() => {
-          deleteBook(realm, item, setList);
+          deleteBook({realm, item, setList, setError});
         }}
       />
     );
   }
 };
 
-const renderItem = (realm, item, setList) => {
+const renderItem = ({realm, item, setList, setError}) => {
   // console.log('item', item);
   if (!item) {
     return null;
@@ -129,7 +142,7 @@ const renderItem = (realm, item, setList) => {
           {item.isbn} {item.isbn13}
         </Text>
       </View>
-      {getIcon(realm, item, setList)}
+      {getIcon({realm, item, setList, setError})}
     </View>
   );
 };
@@ -189,7 +202,7 @@ function BarcodeScanner(props) {
       <View style={styles.listContainer}>
         <FlatList
           data={list}
-          renderItem={({item}) => renderItem(realm, item, setList)}
+          renderItem={({item}) => renderItem({realm, item, setList, setError})}
           keyExtractor={(item, index) => String(index)}
         />
       </View>
